refactor: share the asynchronous retrieve_array_subset_into
`ArrayCached::async_retrieve_array_subset_into_opt` and its
`async_retrieve_multi_chunk_fixed_into` helper were near-verbatim copies of
`Array`'s, differing only in the chunk level retrieval call, so the validation
cascade, the 0/1/many dispatch and the `subdivide` loop all had to be kept in
step by hand.

Factor them into `async_array_read_ops_common`, mirroring what
`array_read_ops_common` already does for the synchronous path. The chunk level
retrieval is supplied by an `AsyncRetrieveInto` implementation rather than the
closures the synchronous helper takes, because the returned futures must be
`Send` on non-`wasm32` targets and the `AsyncFn` traits cannot express that;
`async_trait` boxes them with the `Send`-ness the rest of the asynchronous API
uses.

This also drops the inlined `async_retrieve_chunk_bytes` +
`decode_into_array_bytes_target` in the single-chunk branch of the cached
implementation, which now goes through `async_retrieve_chunk_into` like the
uncached one.
